+++ src/models/Challenge.js
@@ -1,14 +1,22 @@
-'use strict';
-module.exports = (sequelize, DataTypes) => {
-  const challenge = sequelize.define('Challenges', {
-    start: DataTypes.DATE,
-    end: DataTypes.DATE,
-    success: DataTypes.BOOLEAN
-  });
+import Sequelize from "sequelize";
+import { BaseModel, defaultSetting } from "./BaseModel";
+import { BaseChallenge } from "./BaseChallenge";
+import { User } from "./User";
 
-  challenge.associate = function (models) {
-    challenge.belongsTo(models.Users);
-    challenge.belongsTo(models.BaseChallenges);
-  };
-  return challenge;
-};
+export class Challenge extends BaseModel {
+ static load(sequelize) {
+    Challenge.init({
+      start: Sequelize.DATE,
+      end: Sequelize.DATE,
+      success: Sequelize.BOOLEAN
+    }, {
+      ...defaultSetting,
+      sequelize,
+      modelName: 'challenge'
+    })
+  } 
+  static link() {
+    Challenge.belongsTo(User);
+    Challenge.belongsTo(BaseChallenge);
+  }
+}
